requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.18.0
openai>=1.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
